Creates a registration mechanism for snapshot and restore.
